split api client message handling into MessageSender and MessageReceiver
